# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Vehicles Api",
    "description": "Vehicles Api",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/vehicles/": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Vehicles"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "make": {
                  "example": "any"
                },
                "model": {
                  "example": "any"
                },
                "year": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                },
                "type": {
                  "example": "any"
                },
                "mileage": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}